Use references instead of Clone
This removes all `T: Clone` requirements, and instead the Packer will store readonly references. References can be freely copied as needed, so no issues there. I think this is a better approach because the primary usecase of this lib is to later substitute rectangles with some data, like pixels, and cloning that is expensive.`

This PR also:
* bumps the edition to 2021
* adds a few Debug traits
* made `leaf_contains_rect` and `split_tree` not work on `self` due to borrowing rules - easier to separate them. Perhaps `Nodes` could even be made into its own separate struct, as it does not need to reference anything outside of it.